+++ sdk/cosmosdb/cosmos/src/utils/digest.browser.ts
@@ -1,12 +1,14 @@
 import { encodeUTF8 } from "./encode";
 
+const globalThis = typeof self === "undefined" ? window : self;
+
 export async function digest(str: string) {
-  if (!window || !window.crypto || !window.crypto.subtle) {
+  if (!globalThis || !globalThis.crypto || !globalThis.crypto.subtle) {
     throw new Error("Browser does not support cryptography functions");
   }
 
   const data = encodeUTF8(str);
-  const hash = await window.crypto.subtle.digest("SHA-256", data);
+  const hash = await globalThis.crypto.subtle.digest("SHA-256", data);
   return bufferToHex(hash);
 }