+++ sdk/storage/storage-blob/karma.conf.js
@@ -1,7 +1,12 @@
 // https://github.com/karma-runner/karma-chrome-launcher
 process.env.CHROME_BIN = require("puppeteer").executablePath();
 require("dotenv").config({ path: "../.env" });
-const { jsonRecordingFilterFunction, isPlaybackMode } = require("@azure/test-utils-recorder");
+const {
+  jsonRecordingFilterFunction,
+  isPlaybackMode,
+  isSoftRecordMode,
+  isRecordMode
+} = require("@azure/test-utils-recorder");
 
 module.exports = function(config) {
   config.set({
@@ -33,7 +38,7 @@ module.exports = function(config) {
       // Promise,String.prototype.startsWith,String.prototype.endsWith,String.prototype.repeat,String.prototype.includes,Array.prototype.includes,Object.assign,Object.keys
       "https://cdn.polyfill.io/v2/polyfill.js?features=Symbol,Promise,String.prototype.startsWith,String.prototype.endsWith,String.prototype.repeat,String.prototype.includes,Array.prototype.includes,Object.assign,Object.keys|always",
       "dist-test/index.browser.js"
-    ].concat(isPlaybackMode() ? ["recordings/browsers/**/*.json"] : []),
+    ].concat(isPlaybackMode() || isSoftRecordMode() ? ["recordings/browsers/**/*.json"] : []),
 
     // list of files / patterns to exclude
     exclude: [],
@@ -51,7 +56,16 @@ module.exports = function(config) {
     // inject following environment values into browser testing with window.__env__
     // environment values MUST be exported or set with same console running "karma start"
     // https://www.npmjs.com/package/karma-env-preprocessor
-    envPreprocessor: ["ACCOUNT_NAME", "ACCOUNT_SAS", "TEST_MODE", "ACCOUNT_TOKEN"],
+    envPreprocessor: [
+      "ACCOUNT_NAME",
+      "ACCOUNT_SAS",
+      "TEST_MODE",
+      "ACCOUNT_TOKEN",
+      "MD_ACCOUNT_NAME",
+      "MD_ACCOUNT_SAS",
+      "ENCRYPTION_SCOPE_1",
+      "ENCRYPTION_SCOPE_2"
+    ],
 
     // test results reporter to use
     // possible values: 'dots', 'progress'
@@ -69,7 +83,8 @@ module.exports = function(config) {
       reports: {
         lcovonly: "coverage-browser/lcov.info",
         html: "coverage-browser/html/report",
-        "text-summary": null
+        "text-summary": null,
+        cobertura: "./coverage-browser/cobertura-coverage.xml"
       }
     },
 
@@ -119,7 +134,7 @@ module.exports = function(config) {
     browserDisconnectTimeout: 10000,
     browserDisconnectTolerance: 3,
     browserConsoleLogOptions: {
-      terminal: process.env.TEST_MODE !== "record"
+      terminal: !isRecordMode()
     },
 
     client: {