+++ sdk/storage/storage-file-datalake/src/DataLakeServiceClient.ts
@@ -22,7 +22,6 @@ import { createSpan } from "./utils/tracing";
 import { toFileSystemPagedAsyncIterableIterator } from "./transforms";
 import { ServiceGetUserDelegationKeyOptions, ServiceGetUserDelegationKeyResponse } from "./models";
 import { CanonicalCode } from "@opentelemetry/types";
-import { getCachedDefaultHttpClient } from "./utils/cache";
 
 /**
  * DataLakeServiceClient allows you to manipulate Azure
@@ -82,13 +81,6 @@ export class DataLakeServiceClient extends StorageClient {
       | Pipeline,
     options?: StoragePipelineOptions
   ) {
-    // when options.httpClient is not specified, passing in a DefaultHttpClient instance to
-    // avoid each client creating its own http client.
-    const newOptions: StoragePipelineOptions = {
-      httpClient: getCachedDefaultHttpClient(),
-      ...options
-    };
-
     if (credentialOrPipeline instanceof Pipeline) {
       super(url, credentialOrPipeline);
     } else {
@@ -99,7 +91,7 @@ export class DataLakeServiceClient extends StorageClient {
         credential = credentialOrPipeline;
       }
 
-      const pipeline = newPipeline(credential, newOptions);
+      const pipeline = newPipeline(credential, options);
       super(url, pipeline);
     }