+++ sdk/storage/storage-file-share/test/directoryclient.spec.ts
@@ -7,6 +7,8 @@ import { DirectoryCreateResponse } from "../src/generated/src/models";
 import { truncatedISO8061Date } from "../src/utils/utils.common";
 import { TestTracer, setTracer, SpanGraph } from "@azure/core-tracing";
 import { URLBuilder } from "@azure/core-http";
+import { MockPolicyFactory } from "./utils/MockPolicyFactory";
+import { Pipeline } from "../src/Pipeline";
 dotenv.config({ path: "../.env" });
 
 describe("DirectoryClient", () => {
@@ -26,7 +28,7 @@ describe("DirectoryClient", () => {
   fullDirAttributes.notContentIndexed = true;
   fullDirAttributes.noScrubData = true;
 
-  beforeEach(async function() {
+  beforeEach(async function () {
     recorder = record(this, recorderEnvSetup);
     const serviceClient = getBSU();
     shareName = recorder.getUniqueName("share");
@@ -47,7 +49,7 @@ describe("DirectoryClient", () => {
     assert.ok(defaultDirCreateResp.filePermissionKey!);
   });
 
-  afterEach(async function() {
+  afterEach(async function () {
     await shareClient.delete();
     recorder.stop();
   });
@@ -779,7 +781,7 @@ describe("DirectoryClient", () => {
 
     assert.deepStrictEqual(
       await dirClient.forceCloseAllHandles(),
-      { closedHandlesCount: 0 },
+      { closedHandlesCount: 0, closeFailureCount: 0 },
       "Error in forceCloseAllHandles"
     );
   });
@@ -797,6 +799,32 @@ describe("DirectoryClient", () => {
       const handle = result.handleList[0];
       await dirClient.forceCloseHandle(handle.handleId);
     }
+  });
+
+  it("forceCloseHandle could return closeFailureCount", async () => {
+    // TODO: Open or create a handle; currently have to do this manually
+    const result = (
+      await dirClient
+        .listHandles()
+        .byPage()
+        .next()
+    ).value;
+    if (result.handleList !== undefined && result.handleList.length > 0) {
+      const mockPolicyFactory = new MockPolicyFactory({ numberOfHandlesFailedToClose: 1 });
+      const factories = (dirClient as any).pipeline.factories.slice(); // clone factories array
+      factories.unshift(mockPolicyFactory);
+      const pipeline = new Pipeline(factories);
+      const mockDirClient = new ShareDirectoryClient(dirClient.url, pipeline);
+
+      const handle = result.handleList[0];
+      const closeResp = await mockDirClient.forceCloseHandle(handle.handleId);
+      assert.equal(closeResp.closeFailureCount, 1, "Number of handles failed to close is not as set.")
+    }
+  });
+
+  it("forceCloseAllHandles return correct closeFailureCount", async () => {
+    const closeRes = await dirClient.forceCloseAllHandles();
+    assert.equal(closeRes.closeFailureCount, 0, "The closeFailureCount is not set to 0 as default.");
   });
 });