+++ sdk/test-utils/recorder/src/baseRecorder.ts
@@ -10,10 +10,12 @@ import {
   parseUrl,
   isPlaybackMode,
   isRecordMode,
-  findRecordingsFolderPath,
   RecorderEnvironmentSetup,
   filterSecretsFromStrings,
-  filterSecretsRecursivelyFromJSON
+  filterSecretsRecursivelyFromJSON,
+  generateTestRecordingFilePath,
+  nodeRequireRecordingIfExists,
+  windowLens
 } from "./utils";
 import { customConsoleLog } from "./customConsoleLog";
 
@@ -38,6 +40,11 @@ export function setEnvironmentVariables(env: any, replaceableVariables: { [key: 
 export function setEnvironmentOnLoad() {
   if (!isBrowser() && (isRecordMode() || isPlaybackMode())) {
     nock = require("nock");
+    if (!nock.isActive()) {
+      // Nock's restore will also remove the http interceptor itself.
+      // We need to run nock.activate() to re-activate the http interceptor. Without re-activation, nock will not intercept any calls.
+      nock.activate();
+    }
   }
 
   if (isBrowser() && isRecordMode()) {
@@ -55,33 +62,20 @@ export abstract class BaseRecorder {
     customizationsOnRecordings: [],
     queryParametersToSkip: []
   };
+  protected hash: string;
 
-  constructor(platform: "node" | "browsers", testSuiteTitle: string, testTitle: string) {
-    // File Extension
-    // nock recordings for node tests - .js extension
-    // recordings are saved in json format for browser tests - .json extension
-    const ext = platform === "node" ? "js" : "json";
-    // Filepath - `{node|browsers}/<describe-block-title>/recording_<test-title>.{js|json}`
-    this.relativeTestRecordingFilePath =
-      platform +
-      "/" +
-      this.formatPath(testSuiteTitle) +
-      "/recording_" +
-      this.formatPath(testTitle) +
-      "." +
-      ext;
-  }
-
-  protected formatPath(path: string): string {
-    return path
-      .toLowerCase()
-      .replace(/ /g, "_")
-      .replace(/<=/g, "lte")
-      .replace(/>=/g, "gte")
-      .replace(/</g, "lt")
-      .replace(/>/g, "gt")
-      .replace(/=/g, "eq")
-      .replace(/\W/g, "");
+  constructor(
+    platform: "node" | "browsers",
+    hash: string,
+    testSuiteTitle: string,
+    testTitle: string
+  ) {
+    this.hash = hash;
+    this.relativeTestRecordingFilePath = generateTestRecordingFilePath(
+      platform,
+      testSuiteTitle,
+      testTitle
+    );
   }
 
   /**
@@ -117,8 +111,8 @@ export abstract class BaseRecorder {
 }
 
 export class NockRecorder extends BaseRecorder {
-  constructor(testSuiteTitle: string, testTitle: string) {
-    super("node", testSuiteTitle, testTitle);
+  constructor(hash: string, testSuiteTitle: string, testTitle: string) {
+    super("node", hash, testSuiteTitle, testTitle);
   }
 
   public record(recorderEnvironmentSetup: RecorderEnvironmentSetup): void {
@@ -128,7 +122,7 @@ export class NockRecorder extends BaseRecorder {
     });
   }
 
-  public playback(recorderEnvironmentSetup: RecorderEnvironmentSetup, filePath: string): void {
+  public playback(recorderEnvironmentSetup: RecorderEnvironmentSetup, testFilePath: string): void {
     this.environmentSetup = recorderEnvironmentSetup;
     /**
      * `@azure/test-utils-recorder` package is used for both the browser and node tests
@@ -137,70 +131,71 @@ export class NockRecorder extends BaseRecorder {
      *  `path` module is leveraged to import the node test recordings and `path` module can't be imported in the browser.
      *  So, instead of `import`-ing the `path` library, `require` is being used and this code path is never executed in the browser.
      *
-     * [A diiferent strategy is in place to import recordings for browser tests by leveraging `karma` plugins.]
+     * [A different strategy is in place to import recordings for browser tests by leveraging `karma` plugins.]
      */
-    let path = require("path");
-    // Get the full path of the `recordings` folder by navigating through the hierarchy of the test file path.
-    const recordingsFolderPath = findRecordingsFolderPath(filePath);
-    const recordingPath = path.resolve(recordingsFolderPath, this.relativeTestRecordingFilePath);
-    if (fs.existsSync(recordingPath)) {
-      this.uniqueTestInfo = require(recordingPath).testInfo;
-    } else {
-      throw new Error(
-        `Recording (${this.relativeTestRecordingFilePath}) is not found at ${recordingsFolderPath}`
-      );
-    }
+    this.uniqueTestInfo = nodeRequireRecordingIfExists(
+      this.relativeTestRecordingFilePath,
+      testFilePath
+    ).testInfo;
   }
 
   public stop(): void {
-    // Importing "nock" library in the recording and appending the testInfo part in the recording
-    const importNockStatement =
-      "let nock = require('nock');\n" +
-      "\n" +
-      "module.exports.testInfo = " +
-      JSON.stringify(this.uniqueTestInfo) +
-      "\n";
+    if (isRecordMode()) {
+      // Importing "nock" library in the recording and appending the testInfo part in the recording
+      const importNockStatement =
+        "let nock = require('nock');\n" +
+        "\n" +
+        `module.exports.hash = "${this.hash}";\n` +
+        "\n" +
+        "module.exports.testInfo = " +
+        JSON.stringify(this.uniqueTestInfo) +
+        "\n";
 
-    const fixtures = nock.recorder.play();
+      const fixtures = nock.recorder.play();
 
-    // Create the directories recursively incase they don't exist
-    try {
-      // Stripping away the filename from the filepath and retaining the directory structure
-      fs.ensureDirSync(
-        "./recordings/" +
-          this.relativeTestRecordingFilePath.substring(
-            0,
-            this.relativeTestRecordingFilePath.lastIndexOf("/") + 1
-          )
-      );
-    } catch (err) {
-      if (err.code !== "EEXIST") throw err;
+      // Create the directories recursively incase they don't exist
+      try {
+        // Stripping away the filename from the filepath and retaining the directory structure
+        fs.ensureDirSync(
+          "./recordings/" +
+            this.relativeTestRecordingFilePath.substring(
+              0,
+              this.relativeTestRecordingFilePath.lastIndexOf("/") + 1
+            )
+        );
+      } catch (err) {
+        if (err.code !== "EEXIST") throw err;
+      }
+
+      const file = fs.createWriteStream("./recordings/" + this.relativeTestRecordingFilePath, {
+        flags: "w"
+      });
+
+      // Some tests expect errors to happen and, if a writing error is thrown in one of these tests, it may be captured in a catch block by accident,
+      // resulting in unexpected behavior. For this reason we're printing it to the console as well
+      file.on("error", (err: any) => {
+        console.log(err);
+        throw err;
+      });
+
+      file.write(importNockStatement);
+
+      // Saving the recording to the file
+      for (const fixture of fixtures) {
+        // We're not matching query string parameters because they may contain sensitive information, and Nock does not allow us to customize it easily
+        const updatedFixture = fixture.toString().replace(/\.query\(.*\)/, ".query(true)");
+        file.write(this.filterSecrets(updatedFixture) + "\n");
+      }
+
+      file.end();
+
+      nock.recorder.clear();
+      nock.restore();
+    } else if (isPlaybackMode()) {
+      nock.restore();
+      nock.cleanAll();
+      nock.enableNetConnect();
     }
-
-    const file = fs.createWriteStream("./recordings/" + this.relativeTestRecordingFilePath, {
-      flags: "w"
-    });
-
-    // Some tests expect errors to happen and, if a writing error is thrown in one of these tests, it may be captured in a catch block by accident,
-    // resulting in unexpected behavior. For this reason we're printing it to the console as well
-    file.on("error", (err: any) => {
-      console.log(err);
-      throw err;
-    });
-
-    file.write(importNockStatement);
-
-    // Saving the recording to the file
-    for (const fixture of fixtures) {
-      // We're not matching query string parameters because they may contain sensitive information, and Nock does not allow us to customize it easily
-      const updatedFixture = fixture.toString().replace(/\.query\(.*\)/, ".query(true)");
-      file.write(this.filterSecrets(updatedFixture) + "\n");
-    }
-
-    file.end();
-
-    nock.recorder.clear();
-    nock.restore();
   }
 }
 
@@ -220,9 +215,10 @@ export class NockRecorder extends BaseRecorder {
 // This class overrides requests' 'open', 'send' and 'onreadystatechange' functions, adding our own code to them to deal with requests
 export class NiseRecorder extends BaseRecorder {
   private recordings: any[] = [];
+  private xhr: nise.FakeXMLHttpRequestStatic | undefined;
 
-  constructor(testSuiteTitle: string, testTitle: string) {
-    super("browsers", testSuiteTitle, testTitle);
+  constructor(hash: string, testSuiteTitle: string, testTitle: string) {
+    super("browsers", hash, testSuiteTitle, testTitle);
   }
 
   // Inserts a request/response pair into the recordings array
@@ -290,6 +286,7 @@ export class NiseRecorder extends BaseRecorder {
     this.environmentSetup = recorderEnvironmentSetup;
     const self = this;
     const xhr = nise.fakeXhr.useFakeXMLHttpRequest();
+    this.xhr = xhr;
 
     // The following filter allows every request to be sent to the server without being mocked
     xhr.useFilters = true;
@@ -297,7 +294,7 @@ export class NiseRecorder extends BaseRecorder {
 
     // 'onCreate' function is called when a new fake XMLHttpRequest object (req) is created
     // Our intent is to override the request's 'onreadystatechange' function so we can create a recording once the response is ready
-    // We can only override 'onreadystatechange' AFTER the 'send' function is called because we need to make sure our implementation won't be overriden by the client
+    // We can only override 'onreadystatechange' AFTER the 'send' function is called because we need to make sure our implementation won't be overridden by the client
     // But we can only override 'send' AFTER the 'open' function is called because the filter we set above makes Nise override it in 'open' body
     xhr.onCreate = function(req: any) {
       // We'll override the 'open' function, so we need to store a handle to its original implementation
@@ -328,7 +325,7 @@ export class NiseRecorder extends BaseRecorder {
             }
           };
 
-          // Now that we have overriden 'onreadystatechange', we can send the request to the server
+          // Now that we have overridden 'onreadystatechange', we can send the request to the server
           reqSend.apply(req, arguments);
         };
       };
@@ -341,14 +338,19 @@ export class NiseRecorder extends BaseRecorder {
     this.environmentSetup = recorderEnvironmentSetup;
     const self = this;
     const xhr = nise.fakeXhr.useFakeXMLHttpRequest();
+    this.xhr = xhr;
 
     // 'karma-json-preprocessor' helps us to retrieve recordings
-    this.recordings = (window as any).__json__[
-      "recordings/" + this.relativeTestRecordingFilePath
-    ].recordings;
-    this.uniqueTestInfo = (window as any).__json__[
-      "recordings/" + this.relativeTestRecordingFilePath
-    ].uniqueTestInfo;
+    this.recordings = windowLens.get([
+      "__json__",
+      "recordings/" + this.relativeTestRecordingFilePath,
+      "recordings"
+    ]);
+    this.uniqueTestInfo = windowLens.get([
+      "__json__",
+      "recordings/" + this.relativeTestRecordingFilePath,
+      "uniqueTestInfo"
+    ]);
 
     // 'onCreate' function is called when a new fake XMLHttpRequest object (req) is created
     xhr.onCreate = function(req: any) {
@@ -402,19 +404,31 @@ export class NiseRecorder extends BaseRecorder {
   }
 
   public stop(): void {
-    // recordings at this point are in the JSON format.
-    this.recordings = this.filterSecrets(this.recordings);
+    if (isRecordMode()) {
+      // recordings at this point are in the JSON format.
+      this.recordings = this.filterSecrets(this.recordings);
 
-    // We're sending the recordings to the 'karma-json-to-file-reporter' via console.log
-    console.log(
-      JSON.stringify({
-        writeFile: true,
-        path: "./recordings/" + this.relativeTestRecordingFilePath,
-        content: {
-          recordings: this.recordings,
-          uniqueTestInfo: this.uniqueTestInfo
-        }
-      })
-    );
+      // We're sending the recordings to the 'karma-json-to-file-reporter' via console.log
+      console.log(
+        JSON.stringify({
+          writeFile: true,
+          path: "./recordings/" + this.relativeTestRecordingFilePath,
+          content: {
+            recordings: this.recordings,
+            uniqueTestInfo: this.uniqueTestInfo,
+            hash: this.hash
+          }
+        })
+      );
+    } else if (isPlaybackMode()) {
+      // TO DO - playback cleanup if any necessary
+    }
+
+    // Resetting the XHR behavior to it's original state.
+    // Necessary if any code wants to use the browser outside of the recorder once the recorder is stopped.
+    if (this.xhr) {
+      this.xhr.useFilters = false;
+      this.xhr.restore();
+    }
   }
 }