+++ sdk/textanalytics/ai-text-analytics/test/collections.spec.ts
@@ -7,8 +7,9 @@ import { makeAnalyzeSentimentResultCollection } from "../src/analyzeSentimentRes
 import { makeDetectLanguageResultCollection } from "../src/detectLanguageResultCollection";
 import { makeExtractKeyPhrasesResultCollection } from "../src/extractKeyPhrasesResultCollection";
 import { makeRecognizeLinkedEntitiesResultCollection } from "../src/recognizeLinkedEntitiesResultCollection";
-import { makeRecognizeEntitiesResultCollection } from "../src/recognizeEntitiesResultCollection";
+import { makeRecognizeCategorizedEntitiesResultCollection } from "../src/recognizeCategorizedEntitiesResultCollection";
 import { LanguageInput, MultiLanguageInput } from "../src/generated/models";
+import { makeRecognizePiiEntitiesResultCollection } from "../src/recognizePiiEntitiesResultCollection";
 
 describe("SentimentResultCollection", () => {
   it("merges items in order", () => {
@@ -54,7 +55,7 @@ describe("SentimentResultCollection", () => {
         {
           id: "B",
           error: {
-            code: "internalServerError",
+            code: "InternalServerError",
             message: "test error"
           }
         }
@@ -117,7 +118,7 @@ describe("DetectLanguageResultCollection", () => {
         {
           id: "B",
           error: {
-            code: "internalServerError",
+            code: "InternalServerError",
             message: "test error"
           }
         }
@@ -163,7 +164,7 @@ describe("ExtractKeyPhrasesResultCollection", () => {
         {
           id: "B",
           error: {
-            code: "internalServerError",
+            code: "InternalServerError",
             message: "test error"
           }
         }
@@ -177,7 +178,7 @@ describe("ExtractKeyPhrasesResultCollection", () => {
   });
 });
 
-describe("RecognizeEntitiesResultCollection", () => {
+describe("RecognizeCategorizedEntitiesResultCollection", () => {
   it("merges items in order", () => {
     const input: MultiLanguageInput[] = [
       {
@@ -193,7 +194,7 @@ describe("RecognizeEntitiesResultCollection", () => {
         text: "test3"
       }
     ];
-    const result = makeRecognizeEntitiesResultCollection(
+    const result = makeRecognizeCategorizedEntitiesResultCollection(
       input,
       [
         {
@@ -201,7 +202,7 @@ describe("RecognizeEntitiesResultCollection", () => {
           entities: [
             {
               text: "Microsoft",
-              type: "Organization",
+              category: "Organization",
               offset: 10,
               length: 9,
               score: 0.9989
@@ -213,8 +214,8 @@ describe("RecognizeEntitiesResultCollection", () => {
           entities: [
             {
               text: "last week",
-              type: "DateTime",
-              subtype: "DateRange",
+              category: "DateTime",
+              subCategory: "DateRange",
               offset: 34,
               length: 9,
               score: 0.8
@@ -226,7 +227,70 @@ describe("RecognizeEntitiesResultCollection", () => {
         {
           id: "B",
           error: {
-            code: "internalServerError",
+            code: "InternalServerError",
+            message: "test error"
+          }
+        }
+      ],
+      ""
+    );
+
+    const inputOrder = input.map((item) => item.id);
+    const outputOrder = result.map((item) => item.id);
+    assert.deepEqual(inputOrder, outputOrder);
+  });
+});
+
+describe("RecognizePiiEntitiesResultCollection", () => {
+  it("merges items in order", () => {
+    const input: MultiLanguageInput[] = [
+      {
+        id: "A",
+        text: "test"
+      },
+      {
+        id: "B",
+        text: "test2"
+      },
+      {
+        id: "C",
+        text: "test3"
+      }
+    ];
+    const result = makeRecognizePiiEntitiesResultCollection(
+      input,
+      [
+        {
+          id: "A",
+          entities: [
+            {
+              text: "(555) 555-5555",
+              category: "US Phone Number",
+              offset: 10,
+              length: 9,
+              score: 0.9989
+            }
+          ]
+        },
+        {
+          id: "C",
+          entities: [
+            {
+              text: "1234 Default Ln.",
+              category: "US Address",
+              subCategory: "",
+              offset: 34,
+              length: 9,
+              score: 0.8
+            }
+          ]
+        }
+      ],
+      [
+        {
+          id: "B",
+          error: {
+            code: "InternalServerError",
             message: "test error"
           }
         }
@@ -304,7 +368,7 @@ describe("RecognizeLinkedEntitiesResultCollection", () => {
         {
           id: "B",
           error: {
-            code: "internalServerError",
+            code: "InternalServerError",
             message: "test error"
           }
         }