+++ sdk/textanalytics/ai-text-analytics/test/textAnalyticsClient.spec.ts
@@ -59,6 +59,31 @@ describe("[AAD] TextAnalyticsClient", function() {
       assertAllSuccess(results);
     });
 
+    it("service returns error for invalid language", async () => {
+      const [result] = await client.analyzeSentiment(["Hello world!"], "notalanguage");
+      if (result.error === undefined) {
+        assert.fail("Expected an error from the service.");
+        return;
+      }
+      assert.equal(result.error.code, "UnsupportedLanguageCode");
+    });
+
+    it("service returns an error for an empty document", async () => {
+      const data = [...testDataEn];
+      data.splice(1, 0, "");
+      const results = await client.analyzeSentiment(data);
+      const errorResult = results[1];
+      if (errorResult.error === undefined) {
+        assert.fail("Expected an error from the service");
+        return;
+      }
+      assert.equal(
+        results.filter((result) => result.error === undefined).length,
+        testDataEn.length
+      );
+      assert.equal(errorResult.error.code, "InvalidDocument");
+    });
+
     it("client accepts TextDocumentInput[]", async () => {
       const enInputs = testDataEn.map(
         (text): TextDocumentInput => ({
@@ -82,26 +107,31 @@ describe("[AAD] TextAnalyticsClient", function() {
     });
   });
 
-  describe("#detectLanguages", () => {
+  describe("#detectLanguage", () => {
     it("client throws on empty list", async () => {
-      return assert.isRejected(client.detectLanguages([]));
+      return assert.isRejected(client.detectLanguage([]));
     });
 
     it("client accepts no countryHint", async () => {
-      const results = await client.detectLanguages(testDataEn);
+      const results = await client.detectLanguage(testDataEn);
       assert.equal(results.length, testDataEn.length);
       assertAllSuccess(results);
     });
 
     it("client accepts a countryHint", async () => {
-      const results = await client.detectLanguages(["impossible"], "fr");
+      const results = await client.detectLanguage(["impossible"], "fr");
       assert.equal(results.length, 1);
       assertAllSuccess(results);
     });
 
-    it("client produces an error on invalid country hint", async () => {
-      const [result] = await client.detectLanguages(["hello"], "invalidcountry");
-      assert.ok((result as any).error !== undefined);
+    it("service errors on invalid country hint", async () => {
+      const [result] = await client.detectLanguage(["hello"], "invalidcountry");
+      if (result.error === undefined) {
+        assert.fail("Expected an error from the service");
+        return;
+      }
+
+      assert.equal(result.error.code, "InvalidCountryHint");
     });
 
     it("client accepts mixed-country DetectLanguageInput[]", async () => {
@@ -120,7 +150,7 @@ describe("[AAD] TextAnalyticsClient", function() {
       );
       const allInputs = enInputs.concat(esInputs);
 
-      const results = await client.detectLanguages(allInputs);
+      const results = await client.detectLanguage(allInputs);
       assert.equal(results.length, testDataEn.length + testDataEs.length);
       assertAllSuccess(results);
     });
@@ -141,6 +171,20 @@ describe("[AAD] TextAnalyticsClient", function() {
       const results = await client.recognizeEntities(testDataEn, "en");
       assert.equal(results.length, testDataEn.length);
       assertAllSuccess(results);
+    });
+
+    it("service errors on unsupported language", async () => {
+      const [result] = await client.recognizeEntities(
+        ["This is some text, but it doesn't matter."],
+        "notalanguage"
+      );
+
+      if (result.error === undefined) {
+        assert.fail("Expected an error from the service");
+        return;
+      }
+
+      assert.equal(result.error.code, "UnsupportedLanguageCode");
     });
 
     it("client accepts mixed-language TextDocumentInput[]", async () => {
@@ -181,6 +225,20 @@ describe("[AAD] TextAnalyticsClient", function() {
       const results = await client.extractKeyPhrases(testDataEn, "en");
       assert.equal(results.length, testDataEn.length);
       assertAllSuccess(results);
+    });
+
+    it("service errors on unsupported language", async () => {
+      const [result] = await client.extractKeyPhrases(
+        ["This is some text, but it doesn't matter."],
+        "notalanguage"
+      );
+
+      if (result.error === undefined) {
+        assert.fail("Expected an error from the service");
+        return;
+      }
+
+      assert.equal(result.error.code, "UnsupportedLanguageCode");
     });
 
     it("client accepts mixed-language TextDocumentInput[]", async () => {
@@ -236,6 +294,20 @@ describe("[AAD] TextAnalyticsClient", function() {
       }
     });
 
+    it("service errors on unsupported language", async () => {
+      const [result] = await client.recognizePiiEntities(
+        ["This is some text, but it doesn't matter."],
+        "notalanguage"
+      );
+
+      if (result.error === undefined) {
+        assert.fail("Expected an error from the service");
+        return;
+      }
+
+      assert.equal(result.error.code, "UnsupportedLanguageCode");
+    });
+
     it("client accepts mixed-language TextDocumentInput[]", async () => {
       const enInputs = testDataEn.map(
         (text): TextDocumentInput => ({
@@ -275,6 +347,20 @@ describe("[AAD] TextAnalyticsClient", function() {
       const results = await client.recognizeLinkedEntities(testDataEn, "en");
       assert.equal(results.length, testDataEn.length);
       assertAllSuccess(results);
+    });
+
+    it("service errors on unsupported language", async () => {
+      const [result] = await client.recognizeLinkedEntities(
+        ["This is some text, but it doesn't matter."],
+        "notalanguage"
+      );
+
+      if (result.error === undefined) {
+        assert.fail("Expected an error from the service");
+        return;
+      }
+
+      assert.equal(result.error.code, "UnsupportedLanguageCode");
     });
 
     it("client accepts mixed-language TextDocumentInput[]", async () => {