+++ sdk/eventhub/event-hubs/CHANGELOG.md
@@ -1,9 +1,22 @@
 # Release History
 
-## 5.0.1 (TBD)
+## 5.0.2 (Unreleased)
+
+- Fixes an issue that caused `getPartitionIds`, `getEventHubProperties`,
+  and `getPartitionProperties` to throw an error when run against an
+  Event Hub in Azure Stack.
+  ([PR #7463](https://github.com/Azure/azure-sdk-for-js/pull/7463))
+
+## 5.0.1 (2020-02-11)
 
 - Fixed a potential issue with deadlock where greedy consumers could
   starve out other consumers, preventing us from properly balancing.
+- Fixed an issue where calling `subscription.close()` immediately
+  after calling `subscribe` would cause events to still be read.
+- Updated to use the latest version of the `@azure/core-amqp` package which in
+  turn uses the latest version of the `rhea` package.
+  This update improves support for [bundling](https://github.com/Azure/azure-sdk-for-js/blob/master/documentation/Bundling.md) this library.
+  ([Pull Request](https://github.com/amqp/rhea/pull/274))
 
 ## 5.0.0 (2020-01-09)