+++ sdk/storage/storage-file-datalake/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Release History
 
-## 12.0.0-preview.8 (Unreleased)
+## 12.0.0-preview.9 (Unreleased)
 
-- Exposed the interface `OperationTracingOptions` in the API surface to support tracing features. [PR #7174](https://github.com/Azure/azure-sdk-for-js/pull/7174)
+
+## 12.0.0-preview.8 (2020.02)
+
+- Updated Azure Storage Service API version to 2019-07-07.
+- Fixed a bug where the package didn't work as expected when bundling web applications. [PR #7298](https://github.com/Azure/azure-sdk-for-js/pull/7298)
 
 ## 12.0.0-preview.7 (2020.01)