+++ sdk/storage/storage-file-share/CHANGELOG.md
@@ -1,8 +1,15 @@
 # Release History
 
-## 12.0.2 (Unreleased)
+## 12.1.1 (Unreleased)
 
-- Exposed the interface `OperationTracingOptions` in the API surface to support tracing features. [PR #7174](https://github.com/Azure/azure-sdk-for-js/pull/7174)
+
+## 12.1.0 (2020-02-11)
+- Updated Azure Storage Service API version to 2019-07-07.
+- A new type `ShareLeaseClient` is added to manage leases.
+- New SMB parameters for file copy.
+- Force Close Handles Response now includes the number of handles that failed to close.
+- `ShareProperties` now has 4 additional properties for premium file shares.
+- Fixed a bug where the package didn't work as expected when bundling web applications. [PR #7298](https://github.com/Azure/azure-sdk-for-js/pull/7298)
 
 ## 12.0.1 (2020.01)